Source Code for VS Annie Plus
Vs Annie Plus is a version of Atsuover's VS Annie OC Week with a new Main Menu, a new main menu theme and Psych Engine compatibility.

Features 5 Original Songs by Atsuover

-Good Enough

-Lover

-Tug Of War

-Animal

And a Bonus song available in the debug menu.

Out of respect for Atsuover I will not release any assets of VS Annie 2.0 until it is released on Atsuover's GameBanana page.

This mod is best experienced on Hard.